Correct code to follow coding standard
