+++ AI.md
@@ -1,0 +1,2 @@
+Use of AI to generate JavaDoc descriptions for some classes, which saved time
+and reduced tedious work. JavaDocs generated by AI were checked for accuracy.
