+++ src/main/java/yoda/commands/EventCommand.java
@@ -1,12 +1,12 @@
 package yoda.commands;
 
-import yoda.TaskList;
-import yoda.exceptions.YodaException;
-import yoda.tasks.Event;
-
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+
+import yoda.TaskList;
+import yoda.exceptions.YodaException;
+import yoda.tasks.Event;
 
 /**
  * Represents a command to add an event task to the task list.
