+++ src/main/java/yoda/commands/ListCommand.java
@@ -1,9 +1,9 @@
 package yoda.commands;
 
+import java.util.ArrayList;
+
 import yoda.TaskList;
 import yoda.tasks.Task;
-
-import java.util.ArrayList;
 
 /**
  * Represents a command to display list of current tasks.
