--- src/main/java/yoda/exceptions/EmptyInputException.java
@@ -1,7 +1,0 @@
-package yoda.exceptions;
-
-public class EmptyInputException extends YodaException {
-    public EmptyInputException() {
-        super("Hmm... You must speak to be heard...");
-    }
-}
