--- src/main/java/yoda/exceptions/NonsenseInputException.java
@@ -1,7 +1,0 @@
-package yoda.exceptions;
-
-public class NonsenseInputException extends YodaException {
-    public NonsenseInputException() {
-        super("What is the meaning of this...?");
-    }
-}
